Move functions and `payload_dict` out of XenonDevice.py and into new device_helper.py file
