+++ docs/sphinx/_templates/autoclass-module.rst
