+++ tinytuya/core/Device.py
@@ -4,7 +4,8 @@
 import logging
 import time
 
-from .XenonDevice import XenonDevice, merge_dps_results
+from .XenonDevice import XenonDevice
+from .device_helper import merge_dps_results
 from . import command_types as CT
 
 # Fix the name so it's not tinytuya.core.Device
